move mocking of toasts into setupTests
