+++ front/src/views/Box/BoxViewHistoryOverlay.test.tsx
@@ -3,9 +3,6 @@ import { screen, render } from "tests/test-utils";
 import userEvent from "@testing-library/user-event";
 import { cache } from "queries/cache";
 
-import { useErrorHandling } from "hooks/useErrorHandling";
-import { useNotification } from "hooks/useNotification";
-
 import { BOX_BY_LABEL_IDENTIFIER_AND_ALL_SHIPMENTS_QUERY } from "queries/queries";
 import { organisation1 } from "mocks/organisations";
 
@@ -14,13 +11,6 @@ import { history1, history2 } from "mocks/histories";
 import { generateMockBox } from "mocks/boxes";
 import { mockMatchMediaQuery } from "mocks/functions";
 import BTBox from "./BoxView";
-
-const mockedTriggerError = vi.fn();
-const mockedCreateToast = vi.fn();
-vi.mock("hooks/useErrorHandling");
-vi.mock("hooks/useNotification");
-
-cache.reset();
 
 const initialQueryForBoxWithHistory = {
   request: {
@@ -46,10 +36,6 @@ describe("3.1.12 - Box HistoryOverlay on BoxView", () => {
   beforeEach(() => {
     // setting the screensize to
     mockMatchMediaQuery(true);
-    const mockedUseErrorHandling = vi.mocked(useErrorHandling);
-    mockedUseErrorHandling.mockReturnValue({ triggerError: mockedTriggerError });
-    const mockedUseNotification = vi.mocked(useNotification);
-    mockedUseNotification.mockReturnValue({ createToast: mockedCreateToast });
   });
 
   // Test case 3.1.12.1
